requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
langchain==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-chroma==1.1.0
langchain-community==0.3.0
langchain-experimental==0.4.1
chromadb==1.3.7
pypdf==6.4.1
sentence-transformers==5.2.0
aiofiles==24.1.0